Enables Service Principal authentication
Allows the use of Service Principal for authentication to Azure resources.

This change enhances authentication by enabling the use of a Service Principal. It prioritizes SPN credentials if provided, falling back to the default client configuration if SPN details are not available. This provides a more secure and flexible authentication mechanism.